hardwork about decoupling handler.js, fells like use many kinds of my JavaScript & DOM knowledge, but not the accompished edition
