+++ src/scripts/utlis/eventHandler/dbSuccess.js
@@ -1,0 +1,116 @@
+'use strict';
+var dbSuccess = (function dbSuccessGenerator() {
+  var DB = require('indexeddb-crud');
+  var refresh = require('../refresh.js');
+  var createLi = require('../createLi.js');
+  var general = require('./general.js');
+
+  function add() {
+    var inputValue = document.querySelector('#input').value;
+    var list;
+    var newData;
+    var newLi;
+
+    if (inputValue === '') {
+      window.alert('please input a real data~');
+      return 0;
+    }
+    general.ifEmpty.removeInit();
+    newData = _integrateNewData(inputValue);
+    newLi = createLi(newData);
+    list = document.querySelector('#list');
+    list.insertBefore(newLi, list.firstChild); // push newLi to first
+    document.querySelector('#input').value = '';  // reset input's values
+    DB.addItem(newData);
+
+    return 0;
+  }
+
+  function enterAdd(e) {
+    if (e.keyCode === 13) {
+      add();
+    }
+  }
+
+  // li's [x]'s delete
+  function removeLi(e) {
+    var id;
+
+    if (e.target.className === 'close') { // use event delegation
+      // use previously stored data
+      id = parseInt(e.target.parentNode.getAttribute('data-id'), 10);
+      DB.removeItem(id, showAll);
+    }
+  }
+
+  function showInit() {
+    refresh.clear();
+    DB.getAll(refresh.init);
+  }
+
+  function showAll() {
+    refresh.clear();
+    DB.getAll(refresh.all);
+  }
+
+  function showClear() {
+    refresh.clear(); // clear nodes visually
+    refresh.random();
+    DB.clear(); // clear data indeed
+  }
+
+  function showDone() {
+    _showWhetherDone(true);
+  }
+
+  function showTodo() {
+    _showWhetherDone(false);
+  }
+
+  function clickLi(e) {
+    var id;
+    var targetLi = e.target;
+    // use event delegation
+
+    if (targetLi.getAttribute('data-id')) {
+      id = parseInt(targetLi.getAttribute('data-id'), 10); // use previously stored data-id attribute
+      DB.getItem(id, _toggleLi, [targetLi]); // pass _toggleLi and param [e.target] as callback
+    }
+  }
+
+  function _integrateNewData(value) {
+    return {
+      id: DB.getNewKey(),
+      event: value,
+      finished: false,
+      userDate: general.getNewDate('yyyy年MM月dd日 hh:mm')
+    };
+  }
+
+  function _showWhetherDone(whetherDone) {
+    var condition = 'finished';
+
+    refresh.clear();
+    DB.getConditionItem(condition, whetherDone, refresh.part);
+  }
+
+  function _toggleLi(data, targetLi) {
+    targetLi.classList.toggle('finished');
+    data.finished = !data.finished;  // toggle data.finished
+    DB.updateItem(data, showAll);
+  }
+
+  return {
+    add: add,
+    enterAdd: enterAdd,
+    clickLi: clickLi,
+    removeLi: removeLi,
+    showInit: showInit,
+    showAll: showAll,
+    showClear: showClear,
+    showDone: showDone,
+    showTodo: showTodo
+  };
+}());
+
+module.exports = dbSuccess;
