+++ src/scripts/utlis/eventHandler/eventHandler.js
@@ -1,0 +1,12 @@
+'use strict';
+var eventHandler = (function handlerGenerator() {
+  var dbSuccess = require('./dbSuccess.js');
+  var dbFail = require('./dbFail.js');
+
+  return {
+    dbSuccess: dbSuccess,
+    dbFail: dbFail
+  };
+}());
+
+module.exports = eventHandler;
