+++ src/scripts/utlis/eventHandler/general.js
@@ -1,0 +1,58 @@
+var general = (function generalGenerator() {
+  var refresh = require('../refresh.js');
+
+  var ifEmpty = {
+    removeInit: function removeInit() {
+      var list = document.querySelector('#list');
+
+      if (list.firstChild.className === 'aphorism') {
+        list.removeChild(list.firstChild);
+      }
+    },
+    addRandom: function addRandom() {
+      var list = document.querySelector('#list');
+
+      if (list.firstChild) {
+        refresh.random();
+      }
+    }
+  };
+
+
+  // Format date
+  function getNewDate(fmt) {
+    var newDate = new Date();
+    var newfmt = fmt;
+    var o = {
+      'y+': newDate.getFullYear(),
+      'M+': newDate.getMonth() + 1,
+      'd+': newDate.getDate(),
+      'h+': newDate.getHours(),
+      'm+': newDate.getMinutes()
+    };
+    var lens;
+
+    for (var k in o) {
+      if (new RegExp('(' + k + ')').test(newfmt)) {
+        if (k === 'y+') {
+          newfmt = newfmt.replace(RegExp.$1, ('' + o[k]).substr(4 - RegExp.$1.length));
+        } else if (k === 'S+') {
+          lens = RegExp.$1.length;
+          lens = lens === 1 ? 3 : lens;
+          newfmt = newfmt.replace(RegExp.$1, ('00' + o[k]).substr(('' + o[k]).length - 1, lens));
+        } else {
+          newfmt = newfmt.replace(RegExp.$1, (RegExp.$1.length === 1) ? (o[k]) : (('00' + o[k]).substr(('' + o[k]).length)));
+        }
+      }
+    }
+
+    return newfmt;
+  }
+
+  return {
+    ifEmpty: ifEmpty,
+    getNewDate: getNewDate
+  };
+}());
+
+module.exports = general;
